perf-system/requirements.txt
pandas==1.4.4
fastparquet==0.8.3
PyYAML==5.3.1
aiohttp==3.8.1
asyncio==3.4.3
